Refactor and prepare signing
Refactor and prepare signing
