--- .idea/.name
@@ -1,1 +1,0 @@
-Label Ringtone Setter
